app/requirements.txt
Flask==2.2.5
face_recognition
numpy
opencv-python
Pillow
pandas